Move util to block-library package
